requirements.txt
pandas
unidecode
chardet
openpyxl
matplotlib
seaborn
psycopg2-binary
python-decouple
scikit-learn